Added changes to endpoint in delete, now the test passes

+++ backend/controllers/blogs.js
@@ -51,8 +51,10 @@ blogRouter.delete('/:id', async (request, response, next) => {
 
     // Stringify converts the object into a string
     // And slice method removes the double quotes
-    console.log(JSON.stringify(userCreator.user).slice(1, -1) === decodedToken.id)
-
+    const IsTheSameUser = JSON.stringify(userCreator.user).slice(1, -1) === decodedToken.id
+    if (!IsTheSameUser) {
+      return response.status(401).json({ error: 'Only the creator can delete its own notes' });
+    }
     // With this function, it goes, removes it, and returns back that deleted object
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
